pyproject.toml
[tool.poetry]
name = "website"
version = "0.1.0"
description = "Personal website for displaying coding skills"
authors = ["Joe Foster <josephrfoster@protonmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
flask = "=3.0.0"
flask-sqlalchemy = "^3.1.1"
wikipedia = "^1.4.0"
rich = "^13.9.4"
pathlib = "^1.0.1"
requests = "^2.32.3"
pyyaml = "^6.0.2"
markdown = "^3.7"
flask-login = "^0.6.3"
torch = "^2.6.0"
torchvision = "^0.21.0"
flask-scss = "^0.5"
google-cloud-storage = "^3.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
